+++ src/main/java/persistence/Application.java
@@ -5,7 +5,7 @@ import database.H2;
 import jdbc.JdbcTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import persistence.sql.ddl.Person;
+import persistence.domain.Person;
 import persistence.sql.ddl.DdlQuery;
 import persistence.sql.ddl.QueryType;
