+++ src/main/java/persistence/domain/Person.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.domain;
 
 import jakarta.persistence.*;
