+++ src/test/java/persistence/sql/ddl/ColumnQueriesTest.java
@@ -5,6 +5,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import persistence.domain.Person;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
