+++ src/test/java/persistence/sql/ddl/DdlQueryTest.java
@@ -3,12 +3,12 @@ package persistence.sql.ddl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import persistence.domain.Person;
 import persistence.exception.NotEntityException;
 import persistence.study.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("Query 의")
 class DdlQueryTest {
